+++ examples/helpers/movement.rs
@@ -2,6 +2,7 @@ use bevy::{math::Vec3Swizzles, prelude::*};
 use bevy_ecs_tilemap::MapQuery;
 
 #[allow(dead_code)]
+#[derive(Component)]
 pub struct Player;
 
 // A simple camera system for moving and zooming the camera.
