+++ src/lib.rs
@@ -67,7 +67,7 @@ mod map;
 mod map_query;
 mod mesher;
 mod neighbors;
-mod render2;
+mod render;
 mod tile;
 
 pub use crate::chunk::Chunk;
@@ -100,7 +100,7 @@ pub enum IsoType {
 }
 
 /// The type of tile to be rendered, currently we support: Square, Hex, and Isometric.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Component, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum TilemapMeshType {
     Square,
     Hexagon(HexType),
@@ -140,7 +140,7 @@ impl Plugin for TilemapPlugin {
                     .after("hash_update_for_tiles")
                     .after("update_chunk_visibility"),
             )
-            .add_plugin(render2::TilemapRenderPlugin);
+            .add_plugin(render::TilemapRenderPlugin);
     }
 }
 
@@ -177,7 +177,7 @@ pub(crate) fn round_to_power_of_two(value: f32) -> usize {
 }
 
 /// The size of the map, in chunks
-#[derive(Default, Clone, Copy, PartialEq, Eq, Debug)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Eq, Debug)]
 pub struct MapSize(pub u32, pub u32);
 
 impl From<Vec2> for MapSize {
@@ -193,7 +193,7 @@ impl Into<Vec2> for MapSize {
 }
 
 /// The size of each chunk, in tiles
-#[derive(Default, Clone, Copy, PartialEq, Eq, Debug)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Eq, Debug)]
 pub struct ChunkSize(pub u32, pub u32);
 
 impl From<Vec2> for ChunkSize {
@@ -209,7 +209,7 @@ impl Into<Vec2> for ChunkSize {
 }
 
 /// The size of each tile, in pixels
-#[derive(Default, Clone, Copy, PartialEq, Debug)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Debug)]
 pub struct TileSize(pub f32, pub f32);
 
 impl From<Vec2> for TileSize {
@@ -225,7 +225,7 @@ impl Into<Vec2> for TileSize {
 }
 
 /// The size of a texture in pixels
-#[derive(Default, Clone, Copy, PartialEq, Debug)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Debug)]
 pub struct TextureSize(pub f32, pub f32);
 
 impl Into<Vec2> for TextureSize {
@@ -237,7 +237,7 @@ impl Into<Vec2> for TextureSize {
 /// The position of a tile, in map coordinates
 ///
 /// Coordinates start at (0, 0) from the bottom-left tile of the map.
-#[derive(Default, Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Eq, Debug, Hash)]
 pub struct TilePos(pub u32, pub u32);
 
 impl Into<UVec2> for TilePos {
@@ -255,7 +255,7 @@ impl Into<TilePos> for UVec2 {
 /// The position of a tile, in chunk coordinates
 ///
 /// Coordinates start at (0, 0) from the bottom-left tile of the chunk.
-#[derive(Default, Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Eq, Debug, Hash)]
 pub struct LocalTilePos(pub u32, pub u32);
 
 impl Into<UVec2> for LocalTilePos {
@@ -268,7 +268,7 @@ impl Into<UVec2> for LocalTilePos {
 ///
 /// Coordinates start at (0, 0) from the bottom-left chunk of the map.
 /// Note that these coordinates are measured in terms of chunks, not tiles.
-#[derive(Default, Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[derive(Default, Component, Clone, Copy, PartialEq, Eq, Debug, Hash)]
 pub struct ChunkPos(pub u32, pub u32);
 
 impl Into<UVec2> for ChunkPos {
